requirements.txt
Flask
Gunicorn
yfinance
numpy
pandas
python-dotenv